+++ pennylane_qiskit/converter.py
@@ -12,27 +12,85 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-This module contains functions for converting Qiskit QuantumCircuit objects
-into PennyLane circuit templates.
+This module contains functions for converting between Qiskit QuantumCircuit objects
+and PennyLane circuits.
 """
 from typing import Dict, Any
 import warnings
 
 import numpy as np
-from qiskit import QuantumCircuit
+from qiskit import QuantumCircuit, ClassicalRegister, QuantumRegister
+from qiskit import extensions as ex
+from qiskit.converters import circuit_to_dag, dag_to_circuit
+from qiskit.quantum_info import SparsePauliOp
 from qiskit.circuit import Parameter, ParameterExpression
 from qiskit.exceptions import QiskitError
 from sympy import lambdify
 
 import pennylane as qml
 import pennylane.ops as pennylane_ops
-from pennylane_qiskit.qiskit_device import QISKIT_OPERATION_MAP
 
 # pylint: disable=too-many-instance-attributes
+
+QISKIT_OPERATION_MAP_SELF_ADJOINT = {
+    # native PennyLane operations also native to qiskit
+    "PauliX": ex.XGate,
+    "PauliY": ex.YGate,
+    "PauliZ": ex.ZGate,
+    "Hadamard": ex.HGate,
+    "CNOT": ex.CXGate,
+    "CZ": ex.CZGate,
+    "SWAP": ex.SwapGate,
+    "ISWAP": ex.iSwapGate,
+    "RX": ex.RXGate,
+    "RY": ex.RYGate,
+    "RZ": ex.RZGate,
+    "Identity": ex.IGate,
+    "CSWAP": ex.CSwapGate,
+    "CRX": ex.CRXGate,
+    "CRY": ex.CRYGate,
+    "CRZ": ex.CRZGate,
+    "PhaseShift": ex.PhaseGate,
+    "QubitStateVector": ex.Initialize,
+    "StatePrep": ex.Initialize,
+    "Toffoli": ex.CCXGate,
+    "QubitUnitary": ex.UnitaryGate,
+    "U1": ex.U1Gate,
+    "U2": ex.U2Gate,
+    "U3": ex.U3Gate,
+    "IsingZZ": ex.RZZGate,
+    "IsingYY": ex.RYYGate,
+    "IsingXX": ex.RXXGate,
+}
+
+QISKIT_OPERATION_INVERSES_MAP_SELF_ADJOINT = {
+    "Adjoint(" + k + ")": v for k, v in QISKIT_OPERATION_MAP_SELF_ADJOINT.items()
+}
+
+# Separate dictionary for the inverses as the operations dictionary needs
+# to be invertible for the conversion functionality to work
+QISKIT_OPERATION_MAP_NON_SELF_ADJOINT = {"S": ex.SGate, "T": ex.TGate, "SX": ex.SXGate}
+QISKIT_OPERATION_INVERSES_MAP_NON_SELF_ADJOINT = {
+    "Adjoint(S)": ex.SdgGate,
+    "Adjoint(T)": ex.TdgGate,
+    "Adjoint(SX)": ex.SXdgGate,
+}
+
+QISKIT_OPERATION_MAP = {
+    **QISKIT_OPERATION_MAP_SELF_ADJOINT,
+    **QISKIT_OPERATION_MAP_NON_SELF_ADJOINT,
+}
+QISKIT_OPERATION_INVERSES_MAP = {
+    **QISKIT_OPERATION_INVERSES_MAP_SELF_ADJOINT,
+    **QISKIT_OPERATION_INVERSES_MAP_NON_SELF_ADJOINT,
+}
+
+FULL_OPERATION_MAP = {**QISKIT_OPERATION_MAP, **QISKIT_OPERATION_INVERSES_MAP}
 
 inv_map = {v.__name__: k for k, v in QISKIT_OPERATION_MAP.items()}
 
 dagger_map = {"SdgGate": qml.S, "TdgGate": qml.T, "SXdgGate": qml.SX}
+
 
 
 def _check_parameter_bound(param: Parameter, var_ref_map: Dict[Parameter, Any]):
@@ -256,3 +314,108 @@ def load_qasm_from_file(file: str):
         function: the new PennyLane template
     """
     return load(QuantumCircuit.from_qasm_file(file))
+
+def circuit_to_qiskit(circuit, register_size, diagonalize=True, measure=True):
+    """Builds the circuit objects based on the operations and measurements
+    specified to apply.
+
+    Args:
+        operations (list[~.Operation]): operations to apply to the device
+
+    Keyword args:
+        rotations (list[~.Operation]): Operations that rotate the circuit
+            pre-measurement into the eigenbasis of the observables.
+    """
+
+    reg = QuantumRegister(register_size, "q")
+    creg = ClassicalRegister(register_size, "c")
+    qc = QuantumCircuit(register_size, register_size, name="temp")
+
+    for op in circuit.operations:
+        qc &= operation_to_qiskit(op, register_size)
+
+    # rotate the state for measurement in the computational basis
+    if diagonalize:
+        rotations = circuit.diagonalizing_gates
+        for rot in rotations:
+            qc &= operation_to_qiskit(rot, register_size)
+
+    if measure:
+        # barrier ensures we first do all operations, then do all measurements
+        qc.barrier(reg)
+        # we always measure the full register
+        qc.measure(reg, creg)
+
+    return qc
+
+def operation_to_qiskit(operation, register_size=None):
+    """Take a Pennylane operator and convert to a Qiskit circuit
+
+    Args:
+        operation (List[pennylane.Operation]): operation to be converted
+        num_qubits (int): the total number of qubits on the device
+
+    Returns:
+        QuantumCircuit: a quantum circuit objects containing the translated operation
+    """
+    op_wires = operation.wires
+    par = operation.parameters
+
+    # make quantum and classical registers for the full register
+    if register_size is None:
+        register_size = len(op_wires)
+    reg = QuantumRegister(register_size, "q")
+    creg = ClassicalRegister(register_size, "c")
+
+    for idx, p in enumerate(par):
+        if isinstance(p, np.ndarray):
+            # Convert arrays so that Qiskit accepts the parameter
+            par[idx] = p.tolist()
+
+    operation = operation.name
+
+    mapped_operation = FULL_OPERATION_MAP[operation]
+
+    qregs = [reg[i] for i in op_wires.labels]
+
+    adjoint = operation.startswith("Adjoint(")
+    split_op = operation.split("Adjoint(")
+
+    # Need to revert the order of the quantum registers used in
+    # Qiskit such that it matches the PennyLane ordering
+    if adjoint:
+        if split_op[1] in ("QubitUnitary)", "QubitStateVector)", "StatePrep)"):
+            qregs = list(reversed(qregs))
+    else:
+        if split_op[0] in ("QubitUnitary", "QubitStateVector", "StatePrep"):
+            qregs = list(reversed(qregs))
+
+    dag = circuit_to_dag(QuantumCircuit(reg, creg, name=""))
+    gate = mapped_operation(*par)
+
+    dag.apply_operation_back(gate, qargs=qregs)
+    circuit = dag_to_circuit(dag)
+
+    return circuit
+
+def mp_to_pauli(mp, register_size):
+    """Convert a Pauli observable to a SparsePauliOp for measurement via Estimator
+
+        Args:
+            mp(Union[ExpectationMP, VarianceMP]): MeasurementProcess to be converted to a SparsePauliOp
+            register_size(int): total size of the qubit register being measured
+    """
+
+    # ToDo: I believe this could be extended to cover expectation values of Hamiltonians
+
+    observables = {"PauliX": "X",
+                   "PauliY": "Y",
+                   "PauliZ": "Z",
+                   "Identity": "I"}
+
+    pauli_string = ["I"] * register_size
+    pauli_string[mp.wires[0]] = observables[mp.name]
+
+    pauli_string = ('').join(pauli_string)
+
+    return SparsePauliOp(pauli_string)
