+++ pennylane_qiskit/qiskit_device2.py
@@ -24,18 +24,11 @@ from typing import Union, Callable, Tuple, Sequence
 import numpy as np
 import pennylane as qml
 
-
-
-from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
-from qiskit import extensions as ex
 from qiskit.compiler import transpile
-from qiskit.converters import circuit_to_dag, dag_to_circuit
-from qiskit.quantum_info import SparsePauliOp
 
 from qiskit_ibm_runtime import QiskitRuntimeService, Session
 from qiskit_ibm_runtime.constants import RunnerResult
 from qiskit_ibm_runtime import Sampler, Estimator
-
 
 from pennylane import transform
 from pennylane.transforms.core import TransformProgram
@@ -54,65 +47,13 @@ from pennylane.devices.preprocess import (
 from pennylane.measurements import ProbabilityMP, ExpectationMP, VarianceMP
 
 from ._version import __version__
+from .converter import FULL_OPERATION_MAP, circuit_to_qiskit, mp_to_pauli
 
 QuantumTapeBatch = Sequence[QuantumTape]
 QuantumTape_or_Batch = Union[QuantumTape, QuantumTapeBatch]
 Result_or_ResultBatch = Union[Result, ResultBatch]
 
-QISKIT_OPERATION_MAP_SELF_ADJOINT = {
-    # native PennyLane operations also native to qiskit
-    "PauliX": ex.XGate,
-    "PauliY": ex.YGate,
-    "PauliZ": ex.ZGate,
-    "Hadamard": ex.HGate,
-    "CNOT": ex.CXGate,
-    "CZ": ex.CZGate,
-    "SWAP": ex.SwapGate,
-    "ISWAP": ex.iSwapGate,
-    "RX": ex.RXGate,
-    "RY": ex.RYGate,
-    "RZ": ex.RZGate,
-    "Identity": ex.IGate,
-    "CSWAP": ex.CSwapGate,
-    "CRX": ex.CRXGate,
-    "CRY": ex.CRYGate,
-    "CRZ": ex.CRZGate,
-    "PhaseShift": ex.PhaseGate,
-    "QubitStateVector": ex.Initialize,
-    "StatePrep": ex.Initialize,
-    "Toffoli": ex.CCXGate,
-    "QubitUnitary": ex.UnitaryGate,
-    "U1": ex.U1Gate,
-    "U2": ex.U2Gate,
-    "U3": ex.U3Gate,
-    "IsingZZ": ex.RZZGate,
-    "IsingYY": ex.RYYGate,
-    "IsingXX": ex.RXXGate,
-}
 
-QISKIT_OPERATION_INVERSES_MAP_SELF_ADJOINT = {
-    "Adjoint(" + k + ")": v for k, v in QISKIT_OPERATION_MAP_SELF_ADJOINT.items()
-}
-
-# Separate dictionary for the inverses as the operations dictionary needs
-# to be invertible for the conversion functionality to work
-QISKIT_OPERATION_MAP_NON_SELF_ADJOINT = {"S": ex.SGate, "T": ex.TGate, "SX": ex.SXGate}
-QISKIT_OPERATION_INVERSES_MAP_NON_SELF_ADJOINT = {
-    "Adjoint(S)": ex.SdgGate,
-    "Adjoint(T)": ex.TdgGate,
-    "Adjoint(SX)": ex.SXdgGate,
-}
-
-QISKIT_OPERATION_MAP = {
-    **QISKIT_OPERATION_MAP_SELF_ADJOINT,
-    **QISKIT_OPERATION_MAP_NON_SELF_ADJOINT,
-}
-QISKIT_OPERATION_INVERSES_MAP = {
-    **QISKIT_OPERATION_INVERSES_MAP_SELF_ADJOINT,
-    **QISKIT_OPERATION_INVERSES_MAP_NON_SELF_ADJOINT,
-}
-
-FULL_OPERATION_MAP = {**QISKIT_OPERATION_MAP, **QISKIT_OPERATION_INVERSES_MAP}
 
 def accepted_sample_measurement(m: qml.measurements.MeasurementProcess) -> bool:
     """Specifies whether or not a measurement is accepted when sampling."""
@@ -165,115 +106,6 @@ def split_measurement_types(
         return tuple([result[i] for i in sorted(result.keys())])
 
     return tapes, reorder_fn
-
-
-def circuit_to_qiskit(circuit, register_size, diagonalize=True, measure=True):
-    """Builds the circuit objects based on the operations and measurements
-    specified to apply.
-
-    Args:
-        operations (list[~.Operation]): operations to apply to the device
-
-    Keyword args:
-        rotations (list[~.Operation]): Operations that rotate the circuit
-            pre-measurement into the eigenbasis of the observables.
-    """
-
-    reg = QuantumRegister(register_size, "q")
-    creg = ClassicalRegister(register_size, "c")
-    qc = QuantumCircuit(register_size, register_size, name="temp")
-
-    for op in circuit.operations:
-        qc &= operation_to_qiskit(op, register_size)
-
-    # rotate the state for measurement in the computational basis
-    if diagonalize:
-        rotations = circuit.diagonalizing_gates
-        for rot in rotations:
-            qc &= operation_to_qiskit(rot, register_size)
-
-    if measure:
-        # barrier ensures we first do all operations, then do all measurements
-        qc.barrier(reg)
-        # we always measure the full register
-        qc.measure(reg, creg)
-
-    return qc
-
-def operation_to_qiskit(operation, register_size=None):
-    """Take a Pennylane operator and convert to a Qiskit circuit
-
-    Args:
-        operation (List[pennylane.Operation]): operation to be converted
-        num_qubits (int): the total number of qubits on the device
-
-    Returns:
-        QuantumCircuit: a quantum circuit objects containing the translated operation
-    """
-    op_wires = operation.wires
-    par = operation.parameters
-
-    # make quantum and classical registers for the full register
-    if register_size is None:
-        register_size = len(op_wires)
-    reg = QuantumRegister(register_size, "q")
-    creg = ClassicalRegister(register_size, "c")
-
-    for idx, p in enumerate(par):
-        if isinstance(p, np.ndarray):
-            # Convert arrays so that Qiskit accepts the parameter
-            par[idx] = p.tolist()
-
-    operation = operation.name
-
-    mapped_operation = FULL_OPERATION_MAP[operation]
-
-    qregs = [reg[i] for i in op_wires.labels]
-
-    adjoint = operation.startswith("Adjoint(")
-    split_op = operation.split("Adjoint(")
-
-    # Need to revert the order of the quantum registers used in
-    # Qiskit such that it matches the PennyLane ordering
-    if adjoint:
-        if split_op[1] in ("QubitUnitary)", "QubitStateVector)", "StatePrep)"):
-            qregs = list(reversed(qregs))
-    else:
-        if split_op[0] in ("QubitUnitary", "QubitStateVector", "StatePrep"):
-            qregs = list(reversed(qregs))
-
-    dag = circuit_to_dag(QuantumCircuit(reg, creg, name=""))
-    gate = mapped_operation(*par)
-
-    dag.apply_operation_back(gate, qargs=qregs)
-    circuit = dag_to_circuit(dag)
-
-    return circuit
-
-def mp_to_pauli(mp, register_size):
-    """Convert a Pauli observable to a SparsePauliOp for measurement via Estimator
-
-        Args:
-            mp(Union[ExpectationMP, VarianceMP]): MeasurementProcess to be converted to a SparsePauliOp
-            register_size(int): total size of the qubit register being measured
-    """
-
-    # ToDo: I believe this could be extended to cover expectation values of Hamiltonians
-
-    observables = {"PauliX": "X",
-                   "PauliY": "Y",
-                   "PauliZ": "Z",
-                   "Identity": "I"}
-
-    pauli_string = ["I"] * register_size
-    pauli_string[mp.wires[0]] = observables[mp.name]
-
-    pauli_string = ('').join(pauli_string)
-
-    return SparsePauliOp(pauli_string)
-
-
-
 
 
 class QiskitDevice2(Device):
